+++ src/tms9918.h
@@ -61,8 +61,8 @@ void tms9918_init(void);
 
 void tms9918_set_buffer(uint32_t*);
 void tms9918_set_vblint(void (*)(void));
-void tms9918_set_palette(uint8_t);
-void tms9918_set_region(uint8_t);
+void tms9918_set_palette(unsigned);
+void tms9918_set_region(unsigned);
 
 void tms9918_intchk(void);
